refactor: switch to module folder
