+++ zenith/src/lib/modules/launch-list/launch-list-skeleton.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { Skeleton } from '$lib/components/ui/skeleton';
+	import { Skeleton } from '$lib/ui/skeleton';
 </script>
 
 <ul>
