+++ zenith/src/lib/ui/accordion/accordion-content.svelte
