+++ zenith/src/lib/ui/accordion/accordion-item.svelte
