+++ zenith/src/lib/ui/accordion/accordion-trigger.svelte
