+++ zenith/src/lib/ui/context-menu/context-menu-checkbox-item.svelte
