+++ zenith/src/lib/ui/context-menu/context-menu-item.svelte
