+++ zenith/src/lib/ui/context-menu/context-menu-label.svelte
