+++ zenith/src/lib/ui/context-menu/context-menu-shortcut.svelte
