+++ zenith/src/lib/ui/dot/dot.svelte
