+++ zenith/src/lib/ui/terminal/terminal.svelte
